--- app/_recoil/atoms/socialLogin.ts
@@ -1,6 +1,0 @@
-import { atom } from 'recoil'
-
-export const AtomNaverState = atom({
-  key: 'AtomNaverState', // unique ID (with respect to other atoms/selectors)
-  default: '', // default value (aka initial value)
-})
