+++ app/authentication/login/[snsName]/layout.tsx
@@ -1,7 +1,5 @@
-'use client'
-import RecoilProvider from '@/app/_recoil/RecoilProvider'
 import { ReactNode } from 'react'
 
-export default function layout({ children }: { children: ReactNode }) {
-  return <RecoilProvider>{children}</RecoilProvider>
+export default async function layout({ children }: { children: ReactNode }) {
+  return <>{children}</>
 }
