+++ package.json
@@ -16,6 +16,7 @@
     "antd": "^5.13.2",
     "axios": "^1.6.7",
     "classnames": "^2.5.1",
+    "cookies-next": "^4.1.1",
     "next": "14.1.0",
     "react": "^18",
     "react-dom": "^18",
